feat: 게시글 조회, 조회수 구현(엔티티 연관관계 부분 주석처리해둠)

+++ server/src/main/java/fuckingrullet/server/question/dto/QuestionPatchDto.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class QuestionPatchDto {
-    private long questionId;
+    private Integer questionId;
 
     private String title;
     private String article;
